# [11.x] Collection average/avg optimization