--- a/src/Illuminate/Collections/Traits/EnumeratesValues.php
+++ b/src/Illuminate/Collections/Traits/EnumeratesValues.php
@@ -176,6 +176,32 @@ trait EnumeratesValues
     }
 
     /**
+     * Get the average value of a given key.
+     *
+     * @param  (callable(TValue): float|int)|string|null  $callback
+     * @return float|int|null
+     */
+    public function avg($callback = null) {
+        $callback = $this->valueRetriever($callback);
+
+        // Reduce the array to the sum and the count of non-null values.
+        $reduced = $this
+            ->reduce(
+                static function (array &$reduce, mixed $value) use ($callback): array {
+                    if (! is_null($retrieved = $callback($value))) {
+                        $reduce[0] += $retrieved;
+                        $reduce[1]++;
+                    }
+
+                    return $reduce;
+                },
+                [0, 0],
+            );
+
+        return $reduced[1] ? $reduced[0] / $reduced[1] : null;
+    }
+
+    /**
      * Alias for the "avg" method.
      *
      * @param  (callable(TValue): float|int)|string|null  $callback