This PR includes an algorithm improvement for the Collection `avg()` and `average()` function.

The previous code used an inefficient `map+filter+count`, whereas the new implementation does it all at once with a reducer.

### Algo explanation

1. Start with a reducer array with 2 empty elements `[0, 0]`
2. Loop/reduce the array:
   - add (sum) all non-null values to the first array element
   - increment the second array element with 1 (the count)
3. After loop: `$avg = $reduced[0] / $reduced[1]`

Additionally, the function has been moved to the `EnumeratesValues` trait so the lazy collection can also benefit from this optimization. It also works as expected with `yield`.

### Benchmark [(gist)](https://gist.github.com/bert-w/445962b06d1d6abc528824980879fe8f)

| Size     | Old (ms)  | New (ms)   | % Improvement |
|----------|-----------|------------|---------------|
| 10       | 0.0707    | 0.0614     | 13.14%        |
| 100      | 0.3149    | 0.221      | 29.78%        |
| 1,000    | 3.1159    | 2.2254     | 28.57%        |
| 10,000   | 31.0282   | 23.3528    | 24.71%        |
| 100,000  | 315.3512  | 232.7599   | 26.22%        |
| 1,000,000| 3,219.9169| 2,151.9144 | 33.17%        |